+++ fabfile/__init__.py
@@ -8,3 +8,4 @@ import migrate  # noqa
 
 import v_0_1_0  # noqa
 import v_0_1_1  # noqa
+import v_0_2_0  # noqa