+++ iris/service/content/petition/confirmation.py
@@ -8,6 +8,7 @@ from iris.service import rest
 from iris.service.db.dc import dc_update, iso_now_offset
 from iris.service.content.confirmation.handler import Handler
 from iris.service.content.confirmation import Confirmation
+from iris.service.content.user import normalise_phone_number
 from iris.service import sms
 
 from .document import Petition, Supporter
@@ -47,7 +48,7 @@ class SMSBaseHandler(Handler):
             return
         store = False
         user_rel_data = user_rel.relation_dict
-        mobile = user_rel_data.get('mobile')
+        mobile = normalise_phone_number(user_rel_data.get('mobile'))
         if (user.mobile != mobile
             or not user.mobile_trusted
            ):
@@ -114,7 +115,7 @@ class PetitionSMSHandler(SMSBaseHandler, rest.RESTMapper):
         """Confirms the mobile number on the petition
 
         If the mobile number on the owner relation matches the mobile number
-        of this confimation the mobile_trusted flag is set to true.
+        of this confirmation the mobile_trusted flag is set to true.
 
         If the mobile on the relation is the same as the mobile of the
         references user then the users mobile_trusted flag is also set to
@@ -314,7 +315,7 @@ class SupportEMailConfirmHandler(EMailBaseHandler, rest.RESTMapper):
         """Confirms the mobile number on the petition
 
         If the mobile number on the owner relation matches the mobile number
-        of this conrimation the mobile_trusted flag is set to true.
+        of this confirmation the mobile_trusted flag is set to true.
         """
         supporter = self._supporter(confirmation)
         email = supporter.user.relation_dict['email']